requirements.txt
# Core libraries
numpy>=1.24.4
Pillow>=10.0.0
streamlit>=1.26.0
tensorflow>=2.13.0
setuptools>=67.0.0
distlib>=0.3.7
protobuf>=4.24.3


# Additional commonly used libraries
pandas==2.1.1              # For data manipulation
scipy==1.11.2              # For scientific computing
matplotlib==3.8.0          # For data visualization
seaborn==0.12.2            # For statistical data visualization
scikit-learn==1.3.0        # For machine learning utilities
requests==2.31.0           # For making HTTP requests, useful in Streamlit apps
plotly==5.16.0             # For interactive plots in Streamlit
opencv-python-headless==4.8.1.78  # For image processing without GUI support
protobuf==4.24.3           # Required by TensorFlow
